+++ src/pages/GamePage/AutoGameMatchingPage/Auto1vs1/index.tsx
