+++ src/pages/GamePage/AutoGameMatchingPage/AutoCustom/index.tsx
