+++ src/pages/GamePage/AutoGameMatchingPage/AutoCustom/styles.css.ts
@@ -5,7 +5,7 @@ import { rem, theme } from '@/styles';
 export const title = style({
   marginBottom: rem(20),
   color: theme.color.white,
-  fontFamily: 'Tiny5',
+  fontFamily: 'Tiny5, sans-serif',
   fontSize: rem(50),
   textAlign: 'center',
 });
@@ -19,7 +19,7 @@ export const matchArea = style({
 
 export const vs = style({
   color: theme.color.white,
-  fontFamily: 'Tiny5',
+  fontFamily: 'Tiny5, sans-serif',
   fontSize: rem(60),
   marginLeft: rem(80),
   marginRight: rem(80),
