+++ src/pages/GamePage/AutoGameMatchingPage/components/user-card/index.tsx
