+++ src/pages/GamePage/AutoGameMatchingPage/components/user-card/styles.css.ts
