+++ src/pages/GamePage/AutoGameMatchingPage/components/waiting-message/index.tsx
