+++ src/pages/GamePage/AutoGameMatchingPage/index.tsx
