+++ src/pages/GamePage/GameSelectPage/index.tsx
