+++ src/pages/GamePage/GameSelectPage/styles.css.ts
